Tests: Fix PARSEC provider insertion in HttpClientTlsTest
The parsec provider is being inserted at the wrong priority.

The intent is to insert the parsec provider at the head (highest
priority) of the provider list, however insertProviderAt() is 1 based.
Position 0 results in the parsec provider being inserted at the tail
of the list (lowest priority).

Insert the PARSEC at position 1 and verify the position before
continuing.

Signed-off-by: Adam Johnston <adam.johnston@linaro.org>